WsTrust prototype
